allow sub-MHz PCal intervals and propagate vex1.5 P-Cal Base MHz (Wettzell shifted PCal)
git-svn-id: https://svn.atnf.csiro.au/difx@8010 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/difxio/trunk/difxio/difx_input.c
@@ -1337,7 +1337,7 @@ static DifxInput *parseDifxInputDatastreamTable(DifxInput *D, const DifxParamete
 
 			return 0;
 		}
-		D->datastream[e].phaseCalIntervalMHz = atoi(DifxParametersvalue(ip, r));
+		D->datastream[e].phaseCalIntervalMHz = atof(DifxParametersvalue(ip, r));
 
 		r = DifxParametersfind(ip, r+1, "NUM RECORDED FREQS");
 		if(r < 0)
